requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
docopt==0.6.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pipreqs==0.4.11
protobuf==3.20.3
pycodestyle==2.10.0
PyMySQL==1.0.2
python-dotenv==0.21.1
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.0
telebot==0.0.4
typing_extensions==4.4.0
urllib3==1.26.14
yarg==0.1.9
